Use longlong for axis dimensions, size_t for total length

+++ libpigi/gridspec.h
@@ -3,33 +3,36 @@
 #include <hip/hip_runtime.h>
 
 struct GridSpec {
-    size_t Nx;
-    size_t Ny;
+    long long Nx;
+    long long Ny;
     double scalelm;
     double scaleuv;
 
-    static auto fromScaleLM(size_t Nx, size_t Ny, double scalelm) {
+    static auto fromScaleLM(long long Nx, long long Ny, double scalelm) {
         return GridSpec {Nx, Ny, scalelm, 1 / (Nx * scalelm)};
     }
 
-    static auto fromScaleUV(size_t Nx, size_t Ny, double scaleuv) {
+    static auto fromScaleUV(long long Nx, long long Ny, double scaleuv) {
         return GridSpec {Nx, Ny, 1 / (Nx * scaleuv), scaleuv};
     }
 
-    __host__ __device__ inline auto size() const { return Nx * Ny; }
+    __host__ __device__ inline auto size() const {
+        return static_cast<size_t>(Nx) * static_cast<size_t>(Ny);
+    }
 
     __host__ __device__
     inline auto linearToGrid(const size_t idx) const {
         // COLUMN MAJOR ordering
-        auto xpx { static_cast<long long>(idx % Nx) };
-        auto ypx { static_cast<long long>(idx / Nx) };
+        auto xpx { static_cast<long long>(idx) % Nx };
+        auto ypx { static_cast<long long>(idx) / Nx };
         return std::make_tuple(xpx, ypx);
     }
 
     __host__ __device__
-    inline auto gridToLinear(const size_t x, const size_t y) const {
+    inline auto gridToLinear(const long long x, const long long y) const {
         // COLUMN MAJOR ordering
-        return x + y * Nx;
+        return static_cast<size_t>(x) +
+               static_cast<size_t>(y) * static_cast<size_t>(Nx);
     }
 
     template <typename S>
@@ -38,8 +41,8 @@ struct GridSpec {
         auto [lpx, mpx] = linearToGrid(idx);
 
         return std::make_tuple(
-            (lpx - static_cast<long long>(Nx) / 2) * static_cast<S>(scalelm),
-            (mpx - static_cast<long long>(Ny) / 2) * static_cast<S>(scalelm)
+            (lpx - Nx / 2) * static_cast<S>(scalelm),
+            (mpx - Ny / 2) * static_cast<S>(scalelm)
         );
     }
 
@@ -49,24 +52,24 @@ struct GridSpec {
         auto [upx, vpx] = linearToGrid(idx);
 
         return std::make_tuple(
-            (upx - static_cast<long long>(Nx) / 2) * static_cast<S>(scaleuv),
-            (vpx - static_cast<long long>(Ny) / 2) * static_cast<S>(scaleuv)
+            (upx - Nx / 2) * static_cast<S>(scaleuv),
+            (vpx - Ny / 2) * static_cast<S>(scaleuv)
         );
     }
 
     template <typename S>
     inline auto UVtoGrid(S u, S v) {
         return std::make_tuple(
-            static_cast<S>(u / scaleuv + static_cast<long long>(Nx) / 2),
-            static_cast<S>(v / scaleuv + static_cast<long long>(Ny) / 2)
+            static_cast<S>(u / scaleuv + Nx / 2),
+            static_cast<S>(v / scaleuv + Ny / 2)
         );
     }
 
     template <typename S>
     inline auto gridToUV(auto upx, auto vpx) {
         return std::make_tuple(
-            static_cast<S>((upx - static_cast<long long>(Nx) / 2) * scaleuv),
-            static_cast<S>((vpx - static_cast<long long>(Ny) / 2) * scaleuv)
+            static_cast<S>((upx - Nx / 2) * scaleuv),
+            static_cast<S>((vpx - Ny / 2) * scaleuv)
         );
     }
 };
